+++ Covid19Radar/Tests/Covid19Radar.UnitTests/ViewModels/HelpPage/SendLogConfirmationPageViewModelTests.cs
@@ -5,6 +5,7 @@
 using Acr.UserDialogs;
 using Covid19Radar.Services.Logs;
 using Covid19Radar.ViewModels;
+using Covid19Radar.Views;
 using Moq;
 using Prism.Navigation;
 using Xamarin.Forms;
@@ -47,56 +48,6 @@ namespace Covid19Radar.UnitTests.ViewModels
         }
 
         [Fact]
-        public void InitializeTests_CreateZipSuccess()
-        {
-            var testLogId = "test-log-id";
-            mockLogFileService.Setup(x => x.CreateLogId()).Returns(testLogId);
-
-            var testZipFileName = "test-zip-file-name";
-            mockLogFileService.Setup(x => x.CreateZipFileName(testLogId)).Returns(testZipFileName);
-            var testPublicZipFileName = "test-public-zip-file-name";
-            mockLogFileService.Setup(x => x.CreateZipFile(testZipFileName)).Returns(testPublicZipFileName);
-
-            var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
-
-            mockLogFileService.Verify(x => x.CreateLogId(), Times.Once());
-            mockLogFileService.Verify(x => x.CreateZipFileName(testLogId), Times.Once());
-            mockLogFileService.Verify(x => x.CreateZipFile(testZipFileName), Times.Once());
-
-            mockUserDialogs.Verify(x => x.ShowLoading(It.IsAny<string>(), null), Times.Once());
-            mockUserDialogs.Verify(x => x.HideLoading(), Times.Once());
-            mockUserDialogs.Verify(x => x.AlertAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), null), Times.Never());
-
-            mockNavigationService.Verify(x => x.NavigateAsync(It.IsAny<string>()), Times.Never());
-        }
-
-        [Fact]
-        public void InitializeTests_CreateZipFailure()
-        {
-            var testLogId = "test-log-id";
-            mockLogFileService.Setup(x => x.CreateLogId()).Returns(testLogId);
-
-            var testZipFileName = "test-zip-file-name";
-            mockLogFileService.Setup(x => x.CreateZipFileName(testLogId)).Returns(testZipFileName);
-            var testPublicZipFileName = "test-public-zip-file-name";
-            mockLogFileService.Setup(x => x.CreateZipFile(testZipFileName)).Returns(testPublicZipFileName);
-
-            var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
-
-            mockLogFileService.Verify(x => x.CreateLogId(), Times.Once());
-            mockLogFileService.Verify(x => x.CreateZipFileName(testLogId), Times.Once());
-            mockLogFileService.Verify(x => x.CreateZipFile(testZipFileName), Times.Once());
-
-            mockUserDialogs.Verify(x => x.ShowLoading(It.IsAny<string>(), null), Times.Once());
-            mockUserDialogs.Verify(x => x.HideLoading(), Times.Once());
-            mockUserDialogs.Verify(x => x.AlertAsync(It.IsAny<string>(), It.IsAny<string>(), "OK", null), Times.Once());
-
-            mockNavigationService.Verify(x => x.GoBackAsync(), Times.Once());
-        }
-
-        [Fact]
         public void OnClickConfirmLogCommandTests_Success()
         {
             var testLogId = "test-log-id";
@@ -110,7 +61,7 @@ namespace Covid19Radar.UnitTests.ViewModels
             mockLogFileService.Setup(x => x.CopyLogUploadingFileToPublicPath(testZipFileName)).Returns(testPublicZipFilePath);
 
             var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
+            unitUnderTest.Initialize(SendLogConfirmationPage.BuildNavigationParams(testLogId, testPublicZipFilePath));
 
             mockUserDialogs.Invocations.Clear();
             mockLogFileService.Invocations.Clear();
@@ -158,19 +109,19 @@ namespace Covid19Radar.UnitTests.ViewModels
             mockLogFileService.Setup(x => x.DeleteAllLogUploadingFiles()).Returns(true);
 
             var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
+            unitUnderTest.Initialize(SendLogConfirmationPage.BuildNavigationParams(testLogId, testPublicZipFilePath));
 
             mockUserDialogs.Invocations.Clear();
             mockLogFileService.Invocations.Clear();
 
-            mockLogUploadService.Setup(x => x.UploadAsync(testZipFileName)).ReturnsAsync(true);
+            mockLogUploadService.Setup(x => x.UploadAsync(testPublicZipFilePath)).ReturnsAsync(true);
 
             unitUnderTest.OnClickSendLogCommand.Execute(null);
 
             mockUserDialogs.Verify(x => x.ShowLoading(It.IsAny<string>(), null), Times.Once());
             mockUserDialogs.Verify(x => x.HideLoading(), Times.Once());
             mockUserDialogs.Verify(x => x.AlertAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), null), Times.Never());
-            mockLogUploadService.Verify(x => x.UploadAsync(testZipFileName), Times.Once());
+            mockLogUploadService.Verify(x => x.UploadAsync(testPublicZipFilePath), Times.Once());
             mockLogFileService.Verify(x => x.DeleteAllLogUploadingFiles(), Times.Once());
             var expectedParameters = new NavigationParameters { { "logId", testLogId } };
             mockNavigationService.Verify(x => x.NavigateAsync("SendLogCompletePage?useModalNavigation=true/", expectedParameters), Times.Once());
@@ -190,19 +141,19 @@ namespace Covid19Radar.UnitTests.ViewModels
             mockLogFileService.Setup(x => x.CopyLogUploadingFileToPublicPath(testZipFileName)).Returns(testPublicZipFilePath);
 
             var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
+            unitUnderTest.Initialize(SendLogConfirmationPage.BuildNavigationParams(testLogId, testPublicZipFileName));
 
             mockUserDialogs.Invocations.Clear();
             mockLogFileService.Invocations.Clear();
 
-            mockLogUploadService.Setup(x => x.UploadAsync(testZipFileName)).ReturnsAsync(false);
+            mockLogUploadService.Setup(x => x.UploadAsync(testPublicZipFileName)).ReturnsAsync(false);
 
             unitUnderTest.OnClickSendLogCommand.Execute(null);
 
             mockUserDialogs.Verify(x => x.ShowLoading(It.IsAny<string>(), null), Times.Once());
             mockUserDialogs.Verify(x => x.HideLoading(), Times.Once());
             mockUserDialogs.Verify(x => x.AlertAsync(It.IsAny<string>(), It.IsAny<string>(), "OK", null), Times.Once());
-            mockLogUploadService.Verify(x => x.UploadAsync(testZipFileName), Times.Once());
+            mockLogUploadService.Verify(x => x.UploadAsync(testPublicZipFileName), Times.Once());
             mockLogFileService.Verify(x => x.DeleteAllLogUploadingFiles(), Times.Never());
             mockNavigationService.Verify(x => x.NavigateAsync(It.IsAny<string>(), It.IsAny<NavigationParameters>()), Times.Never());
         }
@@ -222,19 +173,21 @@ namespace Covid19Radar.UnitTests.ViewModels
             mockLogFileService.Setup(x => x.DeleteAllLogUploadingFiles()).Returns(false);
 
             var unitUnderTest = CreateViewModel();
-            unitUnderTest.Initialize(new NavigationParameters());
+            unitUnderTest.Initialize(
+                SendLogConfirmationPage.BuildNavigationParams(testLogId, testPublicZipFilePath)
+                );
 
             mockUserDialogs.Invocations.Clear();
             mockLogFileService.Invocations.Clear();
 
-            mockLogUploadService.Setup(x => x.UploadAsync(testZipFileName)).ReturnsAsync(true);
+            mockLogUploadService.Setup(x => x.UploadAsync(testPublicZipFilePath)).ReturnsAsync(true);
 
             unitUnderTest.OnClickSendLogCommand.Execute(null);
 
             mockUserDialogs.Verify(x => x.ShowLoading(It.IsAny<string>(), null), Times.Once());
             mockUserDialogs.Verify(x => x.HideLoading(), Times.Once());
             mockUserDialogs.Verify(x => x.AlertAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), null), Times.Never());
-            mockLogUploadService.Verify(x => x.UploadAsync(testZipFileName), Times.Once());
+            mockLogUploadService.Verify(x => x.UploadAsync(testPublicZipFilePath), Times.Once());
             mockLogFileService.Verify(x => x.DeleteAllLogUploadingFiles(), Times.Once());
             var expectedParameters = new NavigationParameters { { "logId", testLogId } };
             mockNavigationService.Verify(x => x.NavigateAsync("SendLogCompletePage?useModalNavigation=true/", expectedParameters), Times.Once());
